Add Branch and Leaf
Change intersect() Shape op to take a Hit* so an aggregating shape can
return a hit on a child shape.
    Change shape.c
    Change triangle.c
    Change raytrace.c

Add "Shape* ignore" to Hit so that shape deep in a hierarchy can know to
ignore itself.

Add Branch Shape header and source and Jamfile line.

Add Leaf Shape header and source and Jamfile line.